requirements.txt
discord.py>=1.7.0
yfinance>=0.1.55
lxml>=4.6.2
numpy>=1.20.1
matplotlib>=3.3.4
searchtweets>=1.7.6
